fix: rename dir to reflect texting fixtures
